+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+## Bug Fixes
+- Fixed route creation unable to find paths that require passing through a junction, and then re-entering that same junction from a different direction
+
 # v0.1.6 - 2025-10-25
 
 ## Bug Fixes